requirements.txt
annotated-types==0.7.0
anyio==4.9.0
click==8.1.8
cuid2==2.0.1
fastapi==0.115.12
h11==0.14.0
idna==3.10
pydantic==2.11.3
pydantic_core==2.33.1
python-ulid==3.0.0
sniffio==1.3.1
starlette==0.46.1
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.0
httpx==0.28.1
sentient-agent-framework==0.3.0
python-dotenv==1.1.1
eth-abi==5.2.0
eth-utils==5.3.1
pycryptodome==3.23.0
pyyaml==6.0.3
aiohttp==3.13.1
